build: add initial packaging
this adds rpm and debian packaging as well as release generation Makefile.

Like drbd-utils and drbd-reactor, this uses lbvers.py from build-helpers to
check the changelog/spec file.

It has the usual debrelease target, as well as a debrelease-docker target to
generate the release in docker without requiring any npm locally. The idea is
that the "source tarball" already contains the generated dist/ folder that is
then just copied to the right place.

A release would look like this:

- change versions as usual in spec/changelog
- PATH=~/src/linbit/public/build-helpers:$PATH lbvers.py
- PATH=~/src/linbit/public/build-helpers:$PATH make debrelease-docker VERSION=0.1.0 BASE='0.1.0' BUILD='final' BUILD_NR='1' PKG_NR='1'