build: remove version from package.json
When building packages, it makes our lives easier to have a "single
source of truth" regarding the package version.

In our case, this is the "VERSION=$x" we supply when invoking "make".
So, pass this through to the TypeScript code by using a .env file and a
Dotenv Webpack plugin.

This removes one layer of redundancy for the package version.